Add files via upload
So far, I have adder the panels the way I sketched out when I first started the project. I’ve added the prices corresponding to the meals that I plan to have built in to the program. I also added a button for each meal and an add meal button. When you press the add meal button, it adds it to the meal order. I’ve got pictures I want to start adding to the programs within the main working folder.